+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `GET /applicationForm/{applicationFormId}/proposalDataCsv` returns an application form in a usable csv format for making bulk uploads
 - `applicationFormId` is now a field on `bulkUploadTasks`
 
+### Changed
+
+- `ApplicationForms` are no longer created during a bulk upload, and bulk uploads must now explicitly reference an existing application form on creation.
+
 ## 0.28.0 2025-01-08
 
 ### Added